--- src/components/FeaturesSectionHomePage.jsx
@@ -1,132 +1,0 @@
-
-
-export default function FeaturesSectionHomePage() {
-  return (
-   <div>
-     <div> 
-                  <div className="flex justify-center items-center text-center my-10">
-                   <div className="md:w-2/3">
-
-                       <h2 className="font-bold  my-4 text-5xl">Our Features</h2>
-                       <p className="my-4">Activities based on gameplay, environment, or equipment. Examples include team sports, individual sports, water sports, combat sports, motor sports, adventure sports, and esports, promoting competition and skill.</p>
-
-                   </div>
-               </div>
-               </div>
-                
-               <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4">
-                                                                                 <div className=" border  hover:border-gray-600 rounded-lg">
-                                                                                     <figure>
-                                                                                        <img className="rounded-t-lg" src="https://islandpaints.com/wp-content/uploads/2020/09/home-classroom-design-tips-for-online-learning.jpg" alt="" />
-                                                                                     </figure>
-                                                                                     <div className="px-2 my-2">
-                                                                                         <h2 className="font-semibold text-3xl ">Virtual classroom features</h2>
-                                                             
-                                                                                     </div>
-                                                                                     <div className="flex justify-center mb-3">
-                                                                                         <button className="font-bold hover:bg-white px-3 py-2 rounded-lg bg-blue-500 border hover:border-black">See details</button>
-                                                                                     </div>
-                                                             
-                                                             
-                                                                                 </div>
-                                                                                 <div className=" border  hover:border-gray-600 rounded-lg">
-                                                                                     <figure>
-                                                                                        <img className="rounded-t-lg" src="https://islandpaints.com/wp-content/uploads/2020/09/home-classroom-design-tips-for-online-learning.jpg" alt="" />
-                                                                                     </figure>
-                                                                                     <div className="px-2 my-2">
-                                                                                         <h2 className="font-semibold text-3xl ">Virtual classroom features</h2>
-                                                             
-                                                                                     </div>
-                                                                                     <div className="flex justify-center mb-3">
-                                                                                         <button className="font-bold hover:bg-white px-3 py-2 rounded-lg bg-blue-500 border hover:border-black">See details</button>
-                                                                                     </div>
-                                                             
-                                                             
-                                                                                 </div>
-                                                                                 <div className=" border  hover:border-gray-600 rounded-lg">
-                                                                                     <figure>
-                                                                                        <img className="rounded-t-lg" src="https://islandpaints.com/wp-content/uploads/2020/09/home-classroom-design-tips-for-online-learning.jpg" alt="" />
-                                                                                     </figure>
-                                                                                     <div className="px-2 my-2">
-                                                                                         <h2 className="font-semibold text-3xl ">Virtual classroom features</h2>
-                                                             
-                                                                                     </div>
-                                                                                     <div className="flex justify-center mb-3">
-                                                                                         <button className="font-bold hover:bg-white px-3 py-2 rounded-lg bg-blue-500 border hover:border-black">See details</button>
-                                                                                     </div>
-                                                             
-                                                             
-                                                                                 </div>
-                                                                                 <div className=" border  hover:border-gray-600 rounded-lg">
-                                                                                     <figure>
-                                                                                        <img className="rounded-t-lg" src="https://islandpaints.com/wp-content/uploads/2020/09/home-classroom-design-tips-for-online-learning.jpg" alt="" />
-                                                                                     </figure>
-                                                                                     <div className="px-2 my-2">
-                                                                                         <h2 className="font-semibold text-3xl ">Virtual classroom features</h2>
-                                                             
-                                                                                     </div>
-                                                                                     <div className="flex justify-center mb-3">
-                                                                                         <button className="font-bold hover:bg-white px-3 py-2 rounded-lg bg-blue-500 border hover:border-black">See details</button>
-                                                                                     </div>
-                                                             
-                                                             
-                                                                                 </div>
-                                                                                 <div className=" border  hover:border-gray-600 rounded-lg">
-                                                                                     <figure>
-                                                                                        <img className="rounded-t-lg" src="https://islandpaints.com/wp-content/uploads/2020/09/home-classroom-design-tips-for-online-learning.jpg" alt="" />
-                                                                                     </figure>
-                                                                                     <div className="px-2 my-2">
-                                                                                         <h2 className="font-semibold text-3xl ">Virtual classroom features</h2>
-                                                             
-                                                                                     </div>
-                                                                                     <div className="flex justify-center mb-3">
-                                                                                         <button className="font-bold hover:bg-white px-3 py-2 rounded-lg bg-blue-500 border hover:border-black">See details</button>
-                                                                                     </div>
-                                                             
-                                                             
-                                                                                 </div>
-                                                                                 <div className=" border  hover:border-gray-600 rounded-lg">
-                                                                                     <figure>
-                                                                                        <img className="rounded-t-lg" src="https://islandpaints.com/wp-content/uploads/2020/09/home-classroom-design-tips-for-online-learning.jpg" alt="" />
-                                                                                     </figure>
-                                                                                     <div className="px-2 my-2">
-                                                                                         <h2 className="font-semibold text-3xl ">Virtual classroom features</h2>
-                                                             
-                                                                                     </div>
-                                                                                     <div className="flex justify-center mb-3">
-                                                                                         <button className="font-bold hover:bg-white px-3 py-2 rounded-lg bg-blue-500 border hover:border-black">See details</button>
-                                                                                     </div>
-                                                             
-                                                             
-                                                                                 </div><div className=" border  hover:border-gray-600 rounded-lg">
-                                                                                     <figure>
-                                                                                        <img className="rounded-t-lg" src="https://islandpaints.com/wp-content/uploads/2020/09/home-classroom-design-tips-for-online-learning.jpg" alt="" />
-                                                                                     </figure>
-                                                                                     <div className="px-2 my-2">
-                                                                                         <h2 className="font-semibold text-3xl ">Virtual classroom features</h2>
-                                                             
-                                                                                     </div>
-                                                                                     <div className="flex justify-center mb-3">
-                                                                                         <button className="font-bold hover:bg-white px-3 py-2 rounded-lg bg-blue-500 border hover:border-black">See details</button>
-                                                                                     </div>
-                                                             
-                                                             
-                                                                                 </div><div className=" border  hover:border-gray-600 rounded-lg">
-                                                                                     <figure>
-                                                                                        <img className="rounded-t-lg" src="https://islandpaints.com/wp-content/uploads/2020/09/home-classroom-design-tips-for-online-learning.jpg" alt="" />
-                                                                                     </figure>
-                                                                                     <div className="px-2 my-2">
-                                                                                         <h2 className="font-semibold text-3xl ">Virtual classroom features</h2>
-                                                             
-                                                                                     </div>
-                                                                                     <div className="flex justify-center mb-3">
-                                                                                         <button className="font-bold hover:bg-white px-3 py-2 rounded-lg bg-blue-500 border hover:border-black">See details</button>
-                                                                                     </div>
-                                                             
-                                                             
-                                                                                 </div>
-                                                                                 
-                                                                             </div>
-   </div>
-  )
-}
